fix remove accounts

+++ ManageMiniMart/BLL/UserService.cs
@@ -2,6 +2,7 @@
 using ManageMiniMart.DTO;
 using System;
 using System.Collections.Generic;
+using System.Data.Entity.Core.Common.CommandTrees.ExpressionBuilder;
 using System.Data.Entity.Migrations;
 using System.Linq;
 using System.Security.Cryptography;
@@ -44,22 +45,24 @@ namespace ManageMiniMart.BLL
             db.Accounts.AddOrUpdate(account);
             db.SaveChanges();
         }
-        public void removeAccount(List<Account> accounts, Account current)
+        public void removeAccount(List<string> listId, Account current)
         {
             bool isRemove = false;
-            List<Account> accountsToRemove = new List<Account>();
-            foreach (Account account in accounts)
+            foreach(string Id in listId)
             {
-                if (account.person_id != current.person_id)
+                Account account = getAccountByPersonId(Id);
+                if (account != null)
                 {
-                    var shiftWork = db.Shift_work.Where(p => p.person_id == account.person_id).ToList();
-                    db.Shift_work.RemoveRange(shiftWork);
-                    accountsToRemove.Add(account);
-                    isRemove = true;
+                    if (account.person_id != current.person_id)
+                    {
+                        var shiftWork = db.Shift_work.Where(p => p.person_id == account.person_id).ToList();
+                        db.Shift_work.RemoveRange(shiftWork);
+                        db.Accounts.Remove(account);
+                        db.SaveChanges();
+                        isRemove = true;
+                    }
                 }
             }
-            db.Accounts.RemoveRange(accountsToRemove);
-            db.SaveChanges();
             if (!isRemove) throw new Exception("Nothing to remove or you cannot remove your own account");
         }
         public void resetPassword(List<Account> accounts)
